Fix module for puppet 6 support
There are three issues for puppet 6 tests.

The first is that the tdagent provider definition is incorrect. We
don't need to wrap the definition in a parser function as it should
just be loaded.

The second is that we need to include the fluentd when testing the
plugin resource so that the relationship matcher works correctly.

The third is that yumrepo has been moved to it's own module and must be
defined as a dependency.

There is also an issue with the acceptance test because we don't have
the required dependencies to build the test plugin.

Fixes: #39